Move app tool policy tests into connectors
