+++ codex-rs/core/src/connectors_tests.rs
@@ -1,25 +1,13 @@
 use super::*;
 use crate::config::CONFIG_TOML_FILE;
-use crate::config::Config;
 use crate::config::ConfigBuilder;
 use codex_config::AppRequirementToml;
-use codex_config::AppToolRequirementToml;
-use codex_config::AppToolsRequirementsToml;
 use codex_config::AppsRequirementsToml;
 use codex_config::ConfigLayerStack;
 use codex_config::ConfigRequirements;
 use codex_config::ConfigRequirementsToml;
 use codex_config::test_support::CloudConfigBundleFixture;
-use codex_config::types::AppConfig;
-use codex_config::types::AppToolApproval;
-use codex_config::types::AppToolConfig;
-use codex_config::types::AppToolsConfig;
 use codex_config::types::ApprovalsReviewer;
-use codex_config::types::AppsConfigToml;
-use codex_config::types::AppsDefaultConfig;
-use codex_connectors::AppToolPolicy;
-use codex_connectors::AppToolPolicyEvaluator;
-use codex_connectors::AppToolPolicyInput;
 use codex_connectors::merge::plugin_connector_to_app_info;
 use codex_connectors::metadata::connector_install_url;
 use codex_connectors::metadata::sanitize_name;
@@ -27,117 +15,13 @@ use codex_features::Feature;
 use codex_login::CodexAuth;
 use codex_mcp::CODEX_APPS_MCP_SERVER_NAME;
 use codex_mcp::ToolInfo;
-use codex_utils_absolute_path::AbsolutePathBuf;
 use pretty_assertions::assert_eq;
 use rmcp::model::JsonObject;
 use rmcp::model::Tool;
-use rmcp::model::ToolAnnotations;
 use std::collections::BTreeMap;
-use std::collections::HashMap;
 use std::collections::HashSet;
 use std::sync::Arc;
 use tempfile::tempdir;
-
-fn annotations(destructive_hint: Option<bool>, open_world_hint: Option<bool>) -> ToolAnnotations {
-    ToolAnnotations::from_raw(
-        /*title*/ None,
-        /*read_only_hint*/ None,
-        destructive_hint,
-        /*idempotent_hint*/ None,
-        open_world_hint,
-    )
-}
-
-fn app_tool_policy_input<'a>(
-    connector_id: Option<&'a str>,
-    tool_name: &'a str,
-    tool_title: Option<&'a str>,
-    annotations: Option<&ToolAnnotations>,
-) -> AppToolPolicyInput<'a> {
-    AppToolPolicyInput {
-        connector_id,
-        tool_name,
-        tool_title,
-        destructive_hint: annotations.and_then(|annotations| annotations.destructive_hint),
-        open_world_hint: annotations.and_then(|annotations| annotations.open_world_hint),
-    }
-}
-
-fn app_tool_policy_from_apps_config(
-    apps_config: Option<&AppsConfigToml>,
-    connector_id: Option<&str>,
-    tool_name: &str,
-    tool_title: Option<&str>,
-    annotations: Option<&ToolAnnotations>,
-    managed_approval: Option<AppToolApproval>,
-) -> AppToolPolicy {
-    let requirements = managed_approval.map(|approval| {
-        app_tool_requirements(
-            connector_id.expect("managed approval requires a connector id"),
-            tool_name,
-            approval,
-        )
-    });
-    app_tool_policy_from_config_parts(
-        apps_config,
-        requirements.as_ref(),
-        connector_id,
-        tool_name,
-        tool_title,
-        annotations,
-    )
-}
-
-fn app_tool_policy_from_config_parts(
-    apps_config: Option<&AppsConfigToml>,
-    requirements_apps_config: Option<&AppsRequirementsToml>,
-    connector_id: Option<&str>,
-    tool_name: &str,
-    tool_title: Option<&str>,
-    annotations: Option<&ToolAnnotations>,
-) -> AppToolPolicy {
-    let requirements = ConfigRequirementsToml {
-        apps: requirements_apps_config.cloned(),
-        ..Default::default()
-    };
-    let config_layer_stack =
-        ConfigLayerStack::new(Vec::new(), ConfigRequirements::default(), requirements)
-            .expect("config layer stack");
-    let config_layer_stack = if let Some(apps_config) = apps_config {
-        let mut user_config = toml::map::Map::new();
-        user_config.insert(
-            "apps".to_string(),
-            toml::Value::try_from(apps_config).expect("serialize apps config"),
-        );
-        let config_toml_path =
-            AbsolutePathBuf::try_from(std::env::temp_dir().join(CONFIG_TOML_FILE))
-                .expect("absolute config path");
-        config_layer_stack.with_user_config(&config_toml_path, toml::Value::Table(user_config))
-    } else {
-        config_layer_stack
-    };
-    AppToolPolicyEvaluator::new(&config_layer_stack).policy(app_tool_policy_input(
-        connector_id,
-        tool_name,
-        tool_title,
-        annotations,
-    ))
-}
-
-fn app_tool_policy(
-    config: &Config,
-    connector_id: Option<&str>,
-    tool_name: &str,
-    tool_title: Option<&str>,
-    annotations: Option<&ToolAnnotations>,
-) -> AppToolPolicy {
-    AppToolPolicyEvaluator::new(&config.config_layer_stack).policy(app_tool_policy_input(
-        connector_id,
-        tool_name,
-        tool_title,
-        annotations,
-    ))
-}
 
 fn app(id: &str) -> AppInfo {
     AppInfo {
@@ -362,139 +246,6 @@ fn accessible_connectors_from_mcp_tools_preserves_description() {
     );
 }
 
-#[test]
-fn app_tool_policy_uses_global_defaults_for_destructive_hints() {
-    let apps_config = AppsConfigToml {
-        default: Some(AppsDefaultConfig {
-            enabled: true,
-            approvals_reviewer: None,
-            destructive_enabled: false,
-            open_world_enabled: true,
-        }),
-        apps: HashMap::new(),
-    };
-
-    let policy = app_tool_policy_from_apps_config(
-        Some(&apps_config),
-        Some("calendar"),
-        "events/create",
-        /*tool_title*/ None,
-        Some(&annotations(Some(true), /*open_world_hint*/ None)),
-        /*managed_approval*/ None,
-    );
-
-    assert_eq!(
-        policy,
-        AppToolPolicy {
-            enabled: false,
-            approval: AppToolApproval::Auto,
-        }
-    );
-}
-
-#[test]
-fn app_tool_policy_defaults_missing_destructive_hint_to_true() {
-    let apps_config = AppsConfigToml {
-        default: Some(AppsDefaultConfig {
-            enabled: true,
-            approvals_reviewer: None,
-            destructive_enabled: false,
-            open_world_enabled: true,
-        }),
-        apps: HashMap::new(),
-    };
-
-    let policy = app_tool_policy_from_apps_config(
-        Some(&apps_config),
-        Some("calendar"),
-        "events/create",
-        /*tool_title*/ None,
-        Some(&annotations(/*destructive_hint*/ None, Some(false))),
-        /*managed_approval*/ None,
-    );
-
-    assert_eq!(
-        policy,
-        AppToolPolicy {
-            enabled: false,
-            approval: AppToolApproval::Auto,
-        }
-    );
-}
-
-#[test]
-fn app_tool_policy_defaults_missing_open_world_hint_to_true() {
-    let apps_config = AppsConfigToml {
-        default: Some(AppsDefaultConfig {
-            enabled: true,
-            approvals_reviewer: None,
-            destructive_enabled: true,
-            open_world_enabled: false,
-        }),
-        apps: HashMap::new(),
-    };
-
-    let policy = app_tool_policy_from_apps_config(
-        Some(&apps_config),
-        Some("calendar"),
-        "events/create",
-        /*tool_title*/ None,
-        Some(&annotations(Some(false), /*open_world_hint*/ None)),
-        /*managed_approval*/ None,
-    );
-
-    assert_eq!(
-        policy,
-        AppToolPolicy {
-            enabled: false,
-            approval: AppToolApproval::Auto,
-        }
-    );
-}
-
-#[test]
-fn app_is_enabled_uses_default_for_unconfigured_apps() {
-    let apps_config = AppsConfigToml {
-        default: Some(AppsDefaultConfig {
-            enabled: false,
-            approvals_reviewer: None,
-            destructive_enabled: true,
-            open_world_enabled: true,
-        }),
-        apps: HashMap::new(),
-    };
-
-    assert!(!app_is_enabled(&apps_config, Some("calendar")));
-    assert!(!app_is_enabled(&apps_config, /*connector_id*/ None));
-}
-
-#[test]
-fn app_is_enabled_prefers_per_app_override_over_default() {
-    let apps_config = AppsConfigToml {
-        default: Some(AppsDefaultConfig {
-            enabled: false,
-            approvals_reviewer: None,
-            destructive_enabled: true,
-            open_world_enabled: true,
-        }),
-        apps: HashMap::from([(
-            "calendar".to_string(),
-            AppConfig {
-                enabled: true,
-                approvals_reviewer: None,
-                destructive_enabled: None,
-                open_world_enabled: None,
-                default_tools_approval_mode: None,
-                default_tools_enabled: None,
-                tools: None,
-            },
-        )]),
-    };
-
-    assert!(app_is_enabled(&apps_config, Some("calendar")));
-    assert!(!app_is_enabled(&apps_config, Some("drive")));
-}
-
 #[tokio::test]
 async fn app_approvals_reviewer_uses_app_then_default_then_global() {
     for (global, app_default, app, expected_global, expected_default, expected_app) in [
@@ -620,257 +371,6 @@ approvals_reviewer = "user"
     );
 }
 
-#[test]
-fn requirements_disabled_connector_overrides_enabled_connector() {
-    let effective_apps = AppsConfigToml {
-        default: None,
-        apps: HashMap::from([(
-            "connector_123123".to_string(),
-            AppConfig {
-                enabled: true,
-                ..Default::default()
-            },
-        )]),
-    };
-    let requirements_apps = AppsRequirementsToml {
-        apps: BTreeMap::from([(
-            "connector_123123".to_string(),
-            AppRequirementToml {
-                enabled: Some(false),
-                tools: None,
-            },
-        )]),
-    };
-
-    assert_eq!(
-        app_tool_policy_from_config_parts(
-            Some(&effective_apps),
-            Some(&requirements_apps),
-            Some("connector_123123"),
-            "events/list",
-            /*tool_title*/ None,
-            /*annotations*/ None,
-        ),
-        AppToolPolicy {
-            enabled: false,
-            approval: AppToolApproval::Auto,
-        }
-    );
-}
-
-#[test]
-fn requirements_enabled_does_not_override_disabled_connector() {
-    let effective_apps = AppsConfigToml {
-        default: None,
-        apps: HashMap::from([(
-            "connector_123123".to_string(),
-            AppConfig {
-                enabled: false,
-                ..Default::default()
-            },
-        )]),
-    };
-    let requirements_apps = AppsRequirementsToml {
-        apps: BTreeMap::from([(
-            "connector_123123".to_string(),
-            AppRequirementToml {
-                enabled: Some(true),
-                tools: None,
-            },
-        )]),
-    };
-
-    assert_eq!(
-        app_tool_policy_from_config_parts(
-            Some(&effective_apps),
-            Some(&requirements_apps),
-            Some("connector_123123"),
-            "events/list",
-            /*tool_title*/ None,
-            /*annotations*/ None,
-        ),
-        AppToolPolicy {
-            enabled: false,
-            approval: AppToolApproval::Auto,
-        }
-    );
-}
-
-#[tokio::test]
-async fn cloud_config_bundle_disable_connector_overrides_user_apps_config() {
-    let codex_home = tempdir().expect("tempdir should succeed");
-    std::fs::write(
-        codex_home.path().join(CONFIG_TOML_FILE),
-        r#"
-[apps.connector_123123]
-enabled = true
-"#,
-    )
-    .expect("write config");
-
-    let config = ConfigBuilder::default()
-        .codex_home(codex_home.path().to_path_buf())
-        .fallback_cwd(Some(codex_home.path().to_path_buf()))
-        .cloud_config_bundle(
-            CloudConfigBundleFixture::loader_with_enterprise_requirement(
-                r#"
-[apps.connector_123123]
-enabled = false
-"#,
-            ),
-        )
-        .build()
-        .await
-        .expect("config should build");
-
-    let policy = app_tool_policy(
-        &config,
-        Some("connector_123123"),
-        "events.list",
-        /*tool_title*/ None,
-        /*annotations*/ None,
-    );
-    assert_eq!(
-        policy,
-        AppToolPolicy {
-            enabled: false,
-            approval: AppToolApproval::Auto,
-        }
-    );
-}
-
-#[tokio::test]
-async fn cloud_config_bundle_disable_connector_applies_without_user_apps_table() {
-    let codex_home = tempdir().expect("tempdir should succeed");
-    std::fs::write(codex_home.path().join(CONFIG_TOML_FILE), "").expect("write config");
-
-    let config = ConfigBuilder::default()
-        .codex_home(codex_home.path().to_path_buf())
-        .fallback_cwd(Some(codex_home.path().to_path_buf()))
-        .cloud_config_bundle(
-            CloudConfigBundleFixture::loader_with_enterprise_requirement(
-                r#"
-[apps.connector_123123]
-enabled = false
-"#,
-            ),
-        )
-        .build()
-        .await
-        .expect("config should build");
-
-    let policy = app_tool_policy(
-        &config,
-        Some("connector_123123"),
-        "events.list",
-        /*tool_title*/ None,
-        /*annotations*/ None,
-    );
-    assert_eq!(
-        policy,
-        AppToolPolicy {
-            enabled: false,
-            approval: AppToolApproval::Auto,
-        }
-    );
-}
-
-#[tokio::test]
-async fn local_requirements_disable_connector_overrides_user_apps_config() {
-    let codex_home = tempdir().expect("tempdir should succeed");
-    let config_toml_path =
-        AbsolutePathBuf::try_from(codex_home.path().join(CONFIG_TOML_FILE)).expect("abs path");
-    let mut config = ConfigBuilder::default()
-        .codex_home(codex_home.path().to_path_buf())
-        .fallback_cwd(Some(codex_home.path().to_path_buf()))
-        .build()
-        .await
-        .expect("config should build");
-
-    let requirements = ConfigRequirementsToml {
-        apps: Some(AppsRequirementsToml {
-            apps: BTreeMap::from([(
-                "connector_123123".to_string(),
-                AppRequirementToml {
-                    enabled: Some(false),
-                    tools: None,
-                },
-            )]),
-        }),
-        ..Default::default()
-    };
-    config.config_layer_stack =
-        ConfigLayerStack::new(Vec::new(), ConfigRequirements::default(), requirements)
-            .expect("requirements stack")
-            .with_user_config(
-                &config_toml_path,
-                toml::from_str::<toml::Value>(
-                    r#"
-[apps.connector_123123]
-enabled = true
-"#,
-                )
-                .expect("apps config"),
-            );
-
-    let policy = app_tool_policy(
-        &config,
-        Some("connector_123123"),
-        "events.list",
-        /*tool_title*/ None,
-        /*annotations*/ None,
-    );
-    assert_eq!(
-        policy,
-        AppToolPolicy {
-            enabled: false,
-            approval: AppToolApproval::Auto,
-        }
-    );
-}
-
-#[tokio::test]
-async fn local_requirements_disable_connector_applies_without_user_apps_table() {
-    let codex_home = tempdir().expect("tempdir should succeed");
-    let mut config = ConfigBuilder::default()
-        .codex_home(codex_home.path().to_path_buf())
-        .fallback_cwd(Some(codex_home.path().to_path_buf()))
-        .build()
-        .await
-        .expect("config should build");
-
-    let requirements = ConfigRequirementsToml {
-        apps: Some(AppsRequirementsToml {
-            apps: BTreeMap::from([(
-                "connector_123123".to_string(),
-                AppRequirementToml {
-                    enabled: Some(false),
-                    tools: None,
-                },
-            )]),
-        }),
-        ..Default::default()
-    };
-    config.config_layer_stack =
-        ConfigLayerStack::new(Vec::new(), ConfigRequirements::default(), requirements)
-            .expect("requirements stack");
-
-    let policy = app_tool_policy(
-        &config,
-        Some("connector_123123"),
-        "events.list",
-        /*tool_title*/ None,
-        /*annotations*/ None,
-    );
-    assert_eq!(
-        policy,
-        AppToolPolicy {
-            enabled: false,
-            approval: AppToolApproval::Auto,
-        }
-    );
-}
-
 #[tokio::test]
 async fn with_app_enabled_state_preserves_unrelated_disabled_connector() {
     let codex_home = tempdir().expect("tempdir should succeed");
@@ -906,494 +406,6 @@ async fn with_app_enabled_state_preserves_unrelated_disabled_connector() {
     assert_eq!(
         with_app_enabled_state(vec![slack.clone(), app("connector_drive")], &config),
         vec![slack, drive]
-    );
-}
-
-#[test]
-fn app_tool_policy_honors_default_app_enabled_false() {
-    let apps_config = AppsConfigToml {
-        default: Some(AppsDefaultConfig {
-            enabled: false,
-            approvals_reviewer: None,
-            destructive_enabled: true,
-            open_world_enabled: true,
-        }),
-        apps: HashMap::new(),
-    };
-
-    let policy = app_tool_policy_from_apps_config(
-        Some(&apps_config),
-        Some("calendar"),
-        "events/list",
-        /*tool_title*/ None,
-        Some(&annotations(
-            /*destructive_hint*/ None, /*open_world_hint*/ None,
-        )),
-        /*managed_approval*/ None,
-    );
-
-    assert_eq!(
-        policy,
-        AppToolPolicy {
-            enabled: false,
-            approval: AppToolApproval::Auto,
-        }
-    );
-}
-
-#[test]
-fn app_tool_policy_uses_managed_approval_without_apps_config() {
-    let policy = app_tool_policy_from_apps_config(
-        /*apps_config*/ None,
-        Some("calendar"),
-        "events/list",
-        /*tool_title*/ None,
-        /*annotations*/ None,
-        Some(AppToolApproval::Approve),
-    );
-
-    assert_eq!(
-        policy,
-        AppToolPolicy {
-            enabled: true,
-            approval: AppToolApproval::Approve,
-        }
-    );
-}
-
-fn app_tool_requirements(
-    app_id: &str,
-    tool_name: &str,
-    approval_mode: AppToolApproval,
-) -> AppsRequirementsToml {
-    AppsRequirementsToml {
-        apps: BTreeMap::from([(
-            app_id.to_string(),
-            AppRequirementToml {
-                enabled: None,
-                tools: Some(AppToolsRequirementsToml {
-                    tools: BTreeMap::from([(
-                        tool_name.to_string(),
-                        AppToolRequirementToml {
-                            approval_mode: Some(approval_mode),
-                        },
-                    )]),
-                }),
-            },
-        )]),
-    }
-}
-
-#[test]
-fn managed_app_tool_approval_uses_raw_tool_name() {
-    let requirements_apps = app_tool_requirements(
-        "connector_123123",
-        "calendar/list_events",
-        AppToolApproval::Approve,
-    );
-
-    assert_eq!(
-        [
-            app_tool_policy_from_config_parts(
-                /*apps_config*/ None,
-                Some(&requirements_apps),
-                Some("connector_123123"),
-                "calendar/list_events",
-                /*tool_title*/ None,
-                /*annotations*/ None,
-            ),
-            app_tool_policy_from_config_parts(
-                /*apps_config*/ None,
-                Some(&requirements_apps),
-                Some("connector_123123"),
-                "calendar/create_event",
-                /*tool_title*/ None,
-                /*annotations*/ None,
-            ),
-        ],
-        [
-            AppToolPolicy {
-                enabled: true,
-                approval: AppToolApproval::Approve,
-            },
-            AppToolPolicy::default(),
-        ]
-    );
-}
-
-#[tokio::test]
-async fn cloud_config_bundle_tool_approval_overrides_user_apps_config() {
-    let codex_home = tempdir().expect("tempdir should succeed");
-    std::fs::write(
-        codex_home.path().join(CONFIG_TOML_FILE),
-        r#"
-[apps.connector_123123.tools."calendar/list_events"]
-approval_mode = "prompt"
-"#,
-    )
-    .expect("write config");
-
-    let config = ConfigBuilder::default()
-        .codex_home(codex_home.path().to_path_buf())
-        .fallback_cwd(Some(codex_home.path().to_path_buf()))
-        .cloud_config_bundle(
-            CloudConfigBundleFixture::loader_with_enterprise_requirement(
-                r#"
-[apps.connector_123123.tools."calendar/list_events"]
-approval_mode = "approve"
-"#,
-            ),
-        )
-        .build()
-        .await
-        .expect("config should build");
-
-    let policy = app_tool_policy(
-        &config,
-        Some("connector_123123"),
-        "calendar/list_events",
-        /*tool_title*/ None,
-        /*annotations*/ None,
-    );
-    assert_eq!(
-        policy,
-        AppToolPolicy {
-            enabled: true,
-            approval: AppToolApproval::Approve,
-        }
-    );
-}
-
-#[tokio::test]
-async fn local_requirements_tool_approval_overrides_user_apps_config() {
-    let codex_home = tempdir().expect("tempdir should succeed");
-    let config_toml_path =
-        AbsolutePathBuf::try_from(codex_home.path().join(CONFIG_TOML_FILE)).expect("abs path");
-    let mut config = ConfigBuilder::default()
-        .codex_home(codex_home.path().to_path_buf())
-        .fallback_cwd(Some(codex_home.path().to_path_buf()))
-        .build()
-        .await
-        .expect("config should build");
-
-    let requirements = ConfigRequirementsToml {
-        apps: Some(app_tool_requirements(
-            "connector_123123",
-            "calendar/list_events",
-            AppToolApproval::Approve,
-        )),
-        ..Default::default()
-    };
-    config.config_layer_stack =
-        ConfigLayerStack::new(Vec::new(), ConfigRequirements::default(), requirements)
-            .expect("requirements stack")
-            .with_user_config(
-                &config_toml_path,
-                toml::from_str::<toml::Value>(
-                    r#"
-[apps.connector_123123.tools."calendar/list_events"]
-approval_mode = "prompt"
-"#,
-                )
-                .expect("apps config"),
-            );
-
-    let policy = app_tool_policy(
-        &config,
-        Some("connector_123123"),
-        "calendar/list_events",
-        /*tool_title*/ None,
-        /*annotations*/ None,
-    );
-    assert_eq!(
-        policy,
-        AppToolPolicy {
-            enabled: true,
-            approval: AppToolApproval::Approve,
-        }
-    );
-}
-
-#[tokio::test]
-async fn local_requirements_tool_approval_does_not_match_tool_title() {
-    let codex_home = tempdir().expect("tempdir should succeed");
-    let mut config = ConfigBuilder::default()
-        .codex_home(codex_home.path().to_path_buf())
-        .fallback_cwd(Some(codex_home.path().to_path_buf()))
-        .build()
-        .await
-        .expect("config should build");
-
-    let requirements = ConfigRequirementsToml {
-        apps: Some(app_tool_requirements(
-            "connector_123123",
-            "calendar/list_events",
-            AppToolApproval::Approve,
-        )),
-        ..Default::default()
-    };
-    config.config_layer_stack =
-        ConfigLayerStack::new(Vec::new(), ConfigRequirements::default(), requirements)
-            .expect("requirements stack");
-
-    let policy = app_tool_policy(
-        &config,
-        Some("connector_123123"),
-        "calendar/create_event",
-        Some("calendar/list_events"),
-        /*annotations*/ None,
-    );
-    assert_eq!(
-        policy,
-        AppToolPolicy {
-            enabled: true,
-            approval: AppToolApproval::Auto,
-        }
-    );
-}
-
-#[test]
-fn app_tool_policy_allows_per_app_enable_when_default_is_disabled() {
-    let apps_config = AppsConfigToml {
-        default: Some(AppsDefaultConfig {
-            enabled: false,
-            approvals_reviewer: None,
-            destructive_enabled: true,
-            open_world_enabled: true,
-        }),
-        apps: HashMap::from([(
-            "calendar".to_string(),
-            AppConfig {
-                enabled: true,
-                approvals_reviewer: None,
-                destructive_enabled: None,
-                open_world_enabled: None,
-                default_tools_approval_mode: None,
-                default_tools_enabled: None,
-                tools: None,
-            },
-        )]),
-    };
-
-    let policy = app_tool_policy_from_apps_config(
-        Some(&apps_config),
-        Some("calendar"),
-        "events/list",
-        /*tool_title*/ None,
-        Some(&annotations(
-            /*destructive_hint*/ None, /*open_world_hint*/ None,
-        )),
-        /*managed_approval*/ None,
-    );
-
-    assert_eq!(
-        policy,
-        AppToolPolicy {
-            enabled: true,
-            approval: AppToolApproval::Auto,
-        }
-    );
-}
-
-#[test]
-fn app_tool_policy_per_tool_enabled_true_overrides_app_level_disable_flags() {
-    let apps_config = AppsConfigToml {
-        default: None,
-        apps: HashMap::from([(
-            "calendar".to_string(),
-            AppConfig {
-                enabled: true,
-                approvals_reviewer: None,
-                destructive_enabled: Some(false),
-                open_world_enabled: Some(false),
-                default_tools_approval_mode: None,
-                default_tools_enabled: None,
-                tools: Some(AppToolsConfig {
-                    tools: HashMap::from([(
-                        "events/create".to_string(),
-                        AppToolConfig {
-                            enabled: Some(true),
-                            approval_mode: None,
-                        },
-                    )]),
-                }),
-            },
-        )]),
-    };
-
-    let policy = app_tool_policy_from_apps_config(
-        Some(&apps_config),
-        Some("calendar"),
-        "events/create",
-        /*tool_title*/ None,
-        Some(&annotations(Some(true), Some(true))),
-        /*managed_approval*/ None,
-    );
-
-    assert_eq!(
-        policy,
-        AppToolPolicy {
-            enabled: true,
-            approval: AppToolApproval::Auto,
-        }
-    );
-}
-
-#[test]
-fn app_tool_policy_default_tools_enabled_true_overrides_app_level_tool_hints() {
-    let apps_config = AppsConfigToml {
-        default: None,
-        apps: HashMap::from([(
-            "calendar".to_string(),
-            AppConfig {
-                enabled: true,
-                approvals_reviewer: None,
-                destructive_enabled: Some(false),
-                open_world_enabled: Some(false),
-                default_tools_approval_mode: None,
-                default_tools_enabled: Some(true),
-                tools: None,
-            },
-        )]),
-    };
-
-    let policy = app_tool_policy_from_apps_config(
-        Some(&apps_config),
-        Some("calendar"),
-        "events/create",
-        /*tool_title*/ None,
-        Some(&annotations(Some(true), Some(true))),
-        /*managed_approval*/ None,
-    );
-
-    assert_eq!(
-        policy,
-        AppToolPolicy {
-            enabled: true,
-            approval: AppToolApproval::Auto,
-        }
-    );
-}
-
-#[test]
-fn app_tool_policy_default_tools_enabled_false_overrides_app_level_tool_hints() {
-    let apps_config = AppsConfigToml {
-        default: None,
-        apps: HashMap::from([(
-            "calendar".to_string(),
-            AppConfig {
-                enabled: true,
-                approvals_reviewer: None,
-                destructive_enabled: Some(true),
-                open_world_enabled: Some(true),
-                default_tools_approval_mode: Some(AppToolApproval::Approve),
-                default_tools_enabled: Some(false),
-                tools: None,
-            },
-        )]),
-    };
-
-    let policy = app_tool_policy_from_apps_config(
-        Some(&apps_config),
-        Some("calendar"),
-        "events/list",
-        /*tool_title*/ None,
-        Some(&annotations(
-            /*destructive_hint*/ None, /*open_world_hint*/ None,
-        )),
-        /*managed_approval*/ None,
-    );
-
-    assert_eq!(
-        policy,
-        AppToolPolicy {
-            enabled: false,
-            approval: AppToolApproval::Approve,
-        }
-    );
-}
-
-#[test]
-fn app_tool_policy_uses_default_tools_approval_mode() {
-    let apps_config = AppsConfigToml {
-        default: None,
-        apps: HashMap::from([(
-            "calendar".to_string(),
-            AppConfig {
-                enabled: true,
-                approvals_reviewer: None,
-                destructive_enabled: None,
-                open_world_enabled: None,
-                default_tools_approval_mode: Some(AppToolApproval::Prompt),
-                default_tools_enabled: None,
-                tools: Some(AppToolsConfig {
-                    tools: HashMap::new(),
-                }),
-            },
-        )]),
-    };
-
-    let policy = app_tool_policy_from_apps_config(
-        Some(&apps_config),
-        Some("calendar"),
-        "events/list",
-        /*tool_title*/ None,
-        Some(&annotations(
-            /*destructive_hint*/ None, /*open_world_hint*/ None,
-        )),
-        /*managed_approval*/ None,
-    );
-
-    assert_eq!(
-        policy,
-        AppToolPolicy {
-            enabled: true,
-            approval: AppToolApproval::Prompt,
-        }
-    );
-}
-
-#[test]
-fn app_tool_policy_matches_prefix_stripped_tool_name_for_tool_config() {
-    let apps_config = AppsConfigToml {
-        default: None,
-        apps: HashMap::from([(
-            "calendar".to_string(),
-            AppConfig {
-                enabled: true,
-                approvals_reviewer: None,
-                destructive_enabled: Some(false),
-                open_world_enabled: Some(false),
-                default_tools_approval_mode: Some(AppToolApproval::Auto),
-                default_tools_enabled: Some(false),
-                tools: Some(AppToolsConfig {
-                    tools: HashMap::from([(
-                        "events/create".to_string(),
-                        AppToolConfig {
-                            enabled: Some(true),
-                            approval_mode: Some(AppToolApproval::Approve),
-                        },
-                    )]),
-                }),
-            },
-        )]),
-    };
-
-    let policy = app_tool_policy_from_apps_config(
-        Some(&apps_config),
-        Some("calendar"),
-        "calendar_events/create",
-        Some("events/create"),
-        Some(&annotations(Some(true), Some(true))),
-        /*managed_approval*/ None,
-    );
-
-    assert_eq!(
-        policy,
-        AppToolPolicy {
-            enabled: true,
-            approval: AppToolApproval::Approve,
-        }
     );
 }
